wireless-config: fix "use of uninitialized value" warning
When configuring multiple wireless interfaces on the same physical device,
if one of those interfaces is set to type "monitor" and the other to type
"access-point", committing this configuration generates a "use of
uninitialized value" warning.  While this doesn't prevent the above
configuration from committing, it looks better with a clean commit.